requirements.txt
watchdog>=3.0.0
pywin32>=306; sys_platform == 'win32'
psutil>=5.9.0
Pillow>=10.0.0
pystray>=0.19.0
requests>=2.28.0